requirements.txt
docopts
pyyaml
